+++ donaciones-service/src/main/java/ar/edu/utn/frba/ddsi/donaciones/models/entities/donaciones/Donante.java
@@ -1,7 +1,6 @@
-package ar.edu.utn.frba.ddsi.donaciones.models.entities;
+package ar.edu.utn.frba.ddsi.donaciones.models.entities.donaciones;
 
 import ar.edu.utn.frba.ddsi.common.Persona;
-import ar.edu.utn.frba.ddsi.donaciones.models.entities.donaciones.Donacion;
 import java.util.ArrayList;
 import lombok.Data;
