+++ donaciones-service/src/main/java/ar/edu/utn/frba/ddsi/donaciones/models/entities/donaciones/ItemDonacion.java
@@ -1,6 +1,6 @@
 package ar.edu.utn.frba.ddsi.donaciones.models.entities.donaciones;
 
-import ar.edu.utn.frba.ddsi.donaciones.models.entities.Bien;
+import ar.edu.utn.frba.ddsi.donaciones.models.entities.bienes.Bien;
 import lombok.Data;
 
 @Data
