--- donaciones-service/src/main/java/ar/edu/utn/frba/ddsi/donaciones/models/entities/persona/Genero.java
@@ -1,7 +1,0 @@
-package ar.edu.utn.frba.ddsi.donaciones.models.entities.persona;
-
-public enum Genero {
-  HOMBRE,
-  MUJER,
-  OTROS
-}
