--- donaciones-service/src/main/java/ar/edu/utn/frba/ddsi/donaciones/models/entities/persona/medioDeContacto/WhatsApp.java
@@ -1,3 +1,0 @@
-package ar.edu.utn.frba.ddsi.donaciones.models.entities.persona.medioDeContacto;
-
-public class WhatsApp extends Telefono {}
