+++ donaciones-service/src/main/java/ar/edu/utn/frba/ddsi/donaciones/models/entities/personas/Direccion.java
@@ -1,4 +1,4 @@
-package ar.edu.utn.frba.ddsi.donaciones.models.entities.persona;
+package ar.edu.utn.frba.ddsi.donaciones.models.entities.personas;
 
 import lombok.Getter;
 import lombok.Setter;
