+++ donaciones-service/src/main/java/ar/edu/utn/frba/ddsi/donaciones/models/entities/personas/Genero.java
@@ -1,0 +1,7 @@
+package ar.edu.utn.frba.ddsi.donaciones.models.entities.personas;
+
+public enum Genero {
+  HOMBRE,
+  MUJER,
+  OTROS
+}
