+++ donaciones-service/src/main/java/ar/edu/utn/frba/ddsi/donaciones/models/entities/personas/Juridica.java
@@ -1,4 +1,4 @@
-package ar.edu.utn.frba.ddsi.donaciones.models.entities.persona;
+package ar.edu.utn.frba.ddsi.donaciones.models.entities.personas;
 
 import java.util.List;
 import lombok.Getter;
