+++ donaciones-service/src/main/java/ar/edu/utn/frba/ddsi/donaciones/models/entities/personas/medioDeContacto/WhatsApp.java
@@ -1,0 +1,3 @@
+package ar.edu.utn.frba.ddsi.donaciones.models.entities.personas.medioDeContacto;
+
+public class WhatsApp extends Telefono {}
